api/requirements.txt
fastapi==0.61.2
uvicorn==0.12.3
python-dotenv==0.15.0
PyYAML==5.4.1
motor==2.3.1
redis==3.5.3
aio-pika==6.8.0
elasticsearch==7.11.0
elastic-apm==6.0.0
psutil==5.8.0
logzero==1.6.3

#development req
pytest-cov==2.11.1
pytest-bdd==4.0.2